# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Personal Web Backend API",
    "description": "API RESTful para la gestión de artículos y libros del backoffice",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "API Base URL"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Tipo de error",
            "example": "ValidationError"
          },
          "message": {
            "type": "string",
            "description": "Mensaje descriptivo del error",
            "example": "El ID debe ser un UUID v4 válido"
          }
        },
        "required": ["type", "message"]
      },
...
      "Article": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Identificador único del artículo"
          },
          "title": {
            "type": "string",
            "maxLength": 100,
            "description": "Título del artículo"
          },
          "content": {
            "type": "string",
            "maxLength": 10000,
            "description": "Contenido del artículo"
          },
          "excerpt": {
            "type": "string",
            "maxLength": 160,
            "description": "Resumen del artículo"
          },
          "slug": {
            "type": "string",
            "maxLength": 100,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$",
            "description": "URL-friendly identificador"
          },
          "bookIds": {
            "type": "array",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Lista de IDs de libros relacionados"
          },
          "relatedLinks": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "text": {
                  "type": "string",
                  "maxLength": 100,
                  "description": "Texto descriptivo del enlace"
                },
                "url": {
                  "type": "string",
                  "format": "uri",
                  "maxLength": 2000,
                  "description": "URL del enlace"
                }
              },
              "required": ["text", "url"]
            },
            "maxItems": 10,
            "description": "Enlaces relacionados"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de última actualización"
          }
        },
        "required": ["id", "title", "content", "excerpt", "slug", "bookIds", "relatedLinks", "createdAt", "updatedAt"]
      },
      "Book": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Identificador único del libro"
          },
          "title": {
            "type": "string",
            "maxLength": 100,
            "description": "Título del libro"
          },
          "author": {
            "type": "string",
            "maxLength": 100,
            "description": "Autor del libro"
          },
          "isbn": {
            "type": "string",
            "pattern": "^(?=(?:\\D*\\d){10}(?:(?:\\D*\\d){3})?$)[\\d-]+$",
            "description": "ISBN del libro"
          },
          "description": {
            "type": "string",
            "maxLength": 1000,
            "description": "Descripción del libro"
          },
          "purchaseLink": {
            "type": "string",
            "format": "uri",
            "maxLength": 2000,
            "description": "Enlace de compra",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de creación"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Fecha de última actualización"
          }
        },
        "required": ["id", "title", "author", "isbn", "description", "createdAt", "updatedAt"]
      },
      "Collection": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "oneOf": [
                { "$ref": "#/components/schemas/Article" },
                { "$ref": "#/components/schemas/Book" }
              ]
            }
          },
          "page": {
            "type": "integer",
            "minimum": 1,
            "description": "Página actual"
          },
          "limit": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100,
            "description": "Elementos por página"
          },
          "total": {
            "type": "integer",
            "minimum": 0,
            "description": "Total de elementos"
          }
        },
        "required": ["items", "page", "limit", "total"]
      }
    }
  },
  "paths": {
...
    "/backoffice/articles": {
      "get": {
        "summary": "Listar artículos",
        "description": "Obtiene una lista paginada de artículos",
        "tags": ["Articles"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Número de página",
            "schema": {
              "type": "integer",
              "default": 1,
              "minimum": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Número de elementos por página",
            "schema": {
              "type": "integer",
              "default": 10,
              "minimum": 1,
              "maximum": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de artículos obtenida correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Collection"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Crear artículo",
        "description": "Crea un nuevo artículo",
        "tags": ["Articles"],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Article"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Artículo creado correctamente"
          },
          "400": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "No autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "El ID o slug del artículo ya existe",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}